added EscrowFundsDialogForm component, removed useEscrowFundsDialog hook
